memperbaiki bug

+++ app/Models/Item.php
@@ -4,7 +4,7 @@ namespace App\Models;
 
 use Exception;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\BelongsToMany; // Tambahkan ini
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class Item extends Model
@@ -23,22 +23,21 @@ class Item extends Model
         return $this->hasMany(Transaction::class);
     }
 
-    // PENAMBAHAN: Relasi untuk mengambil daftar "resep" (bahan mentah)
     public function bomRawMaterials(): BelongsToMany
     {
         return $this->belongsToMany(Item::class, 'bill_of_materials', 'finished_good_id', 'raw_material_id')
-                    ->withPivot('quantity_required') // Sertakan jumlah yang dibutuhkan
-                    ->withTimestamps();
+                        ->withPivot('quantity_required')
+                        ->withTimestamps();
     }
 
-    public function increaseStock(int $amount): void
+    public function increaseStock(float $amount): void
     {
         $this->quantity += $amount;
         $this->status = $this->quantity > 0 ? 'available' : 'out';
         $this->save();
     }
 
-    public function decreaseStock(int $amount): void
+    public function decreaseStock(float $amount): void
     {
         if ($this->quantity < $amount) {
             throw new Exception('Stok tidak mencukupi untuk transaksi ini.');
